Moving to acoustic wave class
